requirements.txt
beautifulsoup4==4.6.0
cycler==0.10.0
matplotlib==2.0.2
numpy==1.13.1
pyparsing==2.2.0
pypdb==1.103
python-dateutil==2.6.1
pytz==2017.2
six==1.10.0
xmltodict==0.11.0
